Move handler implementation to separate ReadHandler
